+++ src/imio/actionspanel/utils.py
@@ -1,9 +1,6 @@
 # -*- coding: utf-8 -*-
 
-import logging
 from plone import api
-
-logger = logging.getLogger('imio.actionspanel')
 
 
 def unrestrictedRemoveGivenObject(object_to_delete):
@@ -17,11 +14,5 @@ def unrestrictedRemoveGivenObject(object_to_delete):
     """
     # removes the object
     parent = object_to_delete.aq_inner.aq_parent
-    logMsg = '{} at {} deleted by "{}"'.format(
-        object_to_delete.meta_type,
-        object_to_delete.absolute_url_path(),
-        api.user.get_current().getId()
-    )
     with api.env.adopt_roles(['Manager']):
         parent.manage_delObjects(object_to_delete.getId())
-        logger.info(logMsg)